% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{comp_peaks}
\alias{comp_peaks}
\title{Compare Peaks}
\usage{
comp_peaks(ref_peaks, al_peaks)
}
\arguments{
\item{ref_peaks}{a \emph{data.frame} object. A data frame with 4 columns
(Time, X, Y, Peak), ideally the output from either top_peaks() or
thr_peaks().}

\item{al_peaks}{a \emph{data.frame} object. A data frame with 4 columns
(Time, X, Y, Peak), ideally the output from either top_peaks() or
thr_peaks().}
}
\value{
A \emph{data.frame} object. A data frame with 8 columns containing
the matched peaks from the two samples, with the time, x, y, and peak values
for each.
}
\description{
\code{comp_peaks} compares peaks of two samples.
}
\details{
This function find compares the peaks from two samples and
correlates the peaks by determining the peaks closest to each other in the
two samples, within a certain reasonable distance. Then returns a data frame
with a list of the correlated peaks including each of their time coordinates.
}
