% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_fit.R
\name{gauss}
\alias{gauss}
\title{1D Gaussian function}
\usage{
gauss(a, b, c, t)
}
\arguments{
\item{a, b, c}{are \emph{float} objects. Parameters in R^1 for the Gaussian
function.}

\item{t}{a \emph{float} object. The independent variable in R^1 for the
Gaussian function.}
}
\value{
A \emph{float} object. The value of the Gaussian function at time t,
given the parameters input a,b,c.
}
\description{
\code{gauss} Defines the 1D Gaussian curve function.
}
\details{
This function defines a 1D Gaussian curve function.
}
