% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_fit.R
\name{gauss2}
\alias{gauss2}
\title{2D Gaussian function}
\usage{
gauss2(a, b1, b2, c1, c2, t1, t2)
}
\arguments{
\item{a, b1, b2, c1, c2}{are \emph{float} objects. Parameters in R^1 for the
Gaussian function.}

\item{t1, t2}{are \emph{float} objects. The independent variables t=(t1.t2)
in R^2 for the Gaussian function.}
}
\value{
A \emph{float} object. The value of the Gaussian function at time
t=(t1,t2) given the parameters input a,b1,b2,c1,c2.
}
\description{
\code{gauss2} Defines the 2D Gaussian curve function.
}
\details{
This function defines a 2D Gaussian curve function.
}
