% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_nist.R
\name{nist_list}
\alias{nist_list}
\title{Creates list of NIST data}
\usage{
nist_list(nistfile, ...)
}
\arguments{
\item{nistfile}{a \emph{string} object, the file name or path of the MSP file
for the NIST MS Library database.}

\item{...}{additional optional \emph{string} objects, the file names or paths
of the MSP file for the NIST MS Library if the data base is broken into
multiple files.}
}
\value{
nistlist, a \emph{list} object, a list of string vectors for each
compound in the database.
}
\description{
\code{nist_list} creates a list of the data from the NIST MS
database.
}
\details{
This function takes the MSP file containing the data from the NIST
MS Library database and creates a list of string vectors for each compound in
the database.
}
